backend/requirements.txt
Active-Alchemy==1.1.0
annotated-types==0.6.0
anyio==3.7.1
apache-libcloud==3.8.0
arrow==1.3.0
asn1crypto==1.5.1
Assembly==1.3.0
assemblyai==0.20.0
bcrypt==4.1.1
blinker==1.7.0
boto==2.49.0
boto3==1.33.7
botocore==1.33.7
cachelib==0.9.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dicttoxml==1.7.16
distro==1.8.0
exceptiongroup==1.1.3
fastapi==0.104.1
Flask==3.0.0
Flask-Assets==2.1.0
Flask-Caching==2.1.0
Flask-Cloudy==1.1.0
Flask-Cors==4.0.0
Flask-KVSession==0.6.2
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-S3==0.3.3
Flask-SeaSurf==1.1.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.4
inflection==0.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lockfile==0.12.2
Markdown==3.5.1
MarkupSafe==2.1.3
openai==1.3.7
Paginator==0.5.1
passlib==1.7.4
pg8000==1.29.1
pydantic==2.5.1
pydantic_core==2.14.3
PyMySQL==1.1.0
python-dateutil==2.8.0
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.8.2
scramp==1.4.4
SES-Mailer==0.13.0
simplekv==0.14.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
text-unidecode==1.3
tqdm==4.66.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
watchfiles==0.21.0
webassets==2.0
websockets==12.0
Werkzeug==3.0.1
WTForms==3.1.1
